% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fs_poisson}
\alias{fs_poisson}
\title{Models to Simulate Football Match Results}
\usage{
fs_poisson(params, num_teams, games)
}
\arguments{
\item{params}{List. A list of parameters required for the Poisson model}

\item{num_teams}{Integer. The number of teams in the league.}

\item{games}{Data frame. A data frame containing the home and away team indices.
The data frame must have columns: home_index, away_index.}
}
\value{
A list containing the simulated paramaters
}
\description{
Models to Simulate Football Match Results
}
