% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_model_params}
\alias{print_model_params}
\title{Print the required parameters for a specific model}
\usage{
print_model_params(model)
}
\arguments{
\item{model}{A character string with the name of the model (e.g., "poisson").}
}
\value{
None. Prints the expected parameters to the console.
}
\description{
This function prints the parameters expected for each model.
}
