% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{Split Dataset into Training and Testing Sets for Sports Analytics}
\usage{
split_train_test(data, season, rounds_train)
}
\arguments{
\item{data}{A data frame containing sports match data from load_data()}

\item{season}{The specific season to filter data from (must be a character)}

\item{rounds_train}{A vector of rounds to use for training data}
}
\value{
A list containing two data frames: train and test
}
\description{
This function splits a sports dataset into training and testing sets based on
specified season and rounds, with home venue filtering and team indexing.
}
\examples{
\dontrun{
result <- split_train_test(
  data = match_data,
  season = '2022-2023',
  rounds_train = 19
)
train_data <- result$train
test_data <- result$test
}

}
